% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ame_mice.R
\name{ame_mice}
\alias{ame_mice}
\title{ame_mice: estimate average marginal effects using data objects created by mice::mice() (.mids)}
\usage{
ame_mice(model, mice_object, ...)
}
\arguments{
\item{model}{a model object, either lm() or glm(), created using the original data with missing values}

\item{mice_object}{an object of class .mids, created using the mice::mice() function, containing imputed data}

\item{...}{additional arguments (not intended to be used)}
}
\value{
a dataframe of the pooled marginal effect estimates (pooled using Rubin's rules)
}
\description{
ame_mice: estimate average marginal effects using data objects created by mice::mice() (.mids)
}
\examples{
## Create dataset with missingness
data_with_missings <- mtcars |>
 missMethods::delete_MCAR(0.3, "cyl") |>
 dplyr::mutate(cyl = as.factor(cyl))

## Create model object
lm_mod <- lm(formula = mpg ~ wt + hp + cyl + cyl*wt + disp,
             data = data_with_missings)

## Impute data
imputation <- mice::mice(data = data_with_missings, m = 5, seed = 01701)

## Estimate and pool average marginal effects
pooled_mfx <- memice::ame_mice(model = lm_mod, mice_object = imputation)
pooled_mfx
}
