% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_memice.R
\name{initialize_memice}
\alias{initialize_memice}
\title{initialize_memice(): initialize workspace for proper formatting of memice output}
\usage{
initialize_memice()
}
\value{
returns formatting functions in system.frame()
}
\description{
initialize_memice(): initialize workspace for proper formatting of memice output
}
\examples{
memice::initialize_memice()
}
