% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meusv_mice.R
\name{meusv_mice}
\alias{meusv_mice}
\title{meusv_mice: estimate marginal effects at user-specified values using data objects created by mice::mice() (.mids)}
\usage{
meusv_mice(model, mice_object, ...)
}
\arguments{
\item{model}{a model object, either lm() or glm(), created using the original data with missing values}

\item{mice_object}{an object of class .mids, created using the mice::mice() function, containing imputed data}

\item{...}{user-specified values}
}
\value{
a dataframe of the pooled marginal effect estimates (pooled using Rubin's rules)
}
\description{
meusv_mice: estimate marginal effects at user-specified values using data objects created by mice::mice() (.mids)
}
\examples{
## generate random data for example
## Create dataset with missingness
data_with_missings <- mtcars |>
 missMethods::delete_MCAR(0.3, "cyl") |>
 dplyr::mutate(cyl = as.factor(cyl))

## Create model object
lm_mod <- lm(formula = mpg ~ wt + hp + cyl + cyl*wt + disp,
             data = data_with_missings)

## Impute data
imputation <- mice::mice(data = data_with_missings, m = 5, seed = 01701)

## Estimate and pool marginal effects at user-specified values
memice::initialize_memice()
pooled_mfx <- memice::meusv_mice(model = lm_mod, mice_object = imputation, cyl = 4, wt = 3.5)
pooled_mfx
}
